Clean dashboard component
